Draft LaTeX Code
Here is my first few items in the LaTeX code.  I was not able to get to 
all of it just yet, but I wanted you to know that I am working on it and 
I will send more later tonight.  Here is my first commit to the project 
in a separate branch.